feat: 패시브아이템데이터 추가
-데이터값추가
-엑셀so제너레이터 수정

+++ Assets/02.Scripts/Editor/ExcelSOGenerator.cs
@@ -106,7 +106,7 @@ public static class ExcelSOGenerator
                     itemData.name = row[1].ToString();
                     itemData.Rarity = (Rarity)int.Parse(row[2].ToString());
                     itemData.ConditionType = ParseEnumFromCell<ConditionType>(row[3]);
-                    itemData.value = int.Parse(row[4].ToString());
+                    itemData.value = float.Parse(row[4].ToString());
                     itemData.description = row[5].ToString().Replace("@", itemData.value.ToString());
                     itemData.IconPath = row[6].ToString();
                     itemData.buyPrice = int.Parse(row[7].ToString());
